+++ clang-tools-extra/clang-tidy/openmp/OpenMPTidyModule.cpp
@@ -9,6 +9,7 @@
 #include "../ClangTidy.h"
 #include "../ClangTidyModule.h"
 #include "../ClangTidyModuleRegistry.h"
+#include "AvoidNestingCriticalSectionsCheck.h"
 #include "DeclareLoopIterationVariableInForInitStatementCheck.h"
 #include "DoNotModifyLoopVariableCheck.h"
 #include "ExceptionEscapeCheck.h"
@@ -24,6 +25,8 @@ namespace openmp {
 class OpenMPModule : public ClangTidyModule {
 public:
   void addCheckFactories(ClangTidyCheckFactories &CheckFactories) override {
+    CheckFactories.registerCheck<AvoidNestingCriticalSectionsCheck>(
+        "openmp-avoid-nesting-critical-sections");
     CheckFactories
         .registerCheck<DeclareLoopIterationVariableInForInitStatementCheck>(
             "openmp-declare-loop-iteration-variable-in-for-init-statement");